Add workaround in log4j legacy API test
log4j doesn't seem to make it possible clear loggers the way
JUL does, so all loggers defined in previous tests are still
present in subsequent ones.

This coupled with the fact that the legacy agent attaches one
handler to the root logger, all loggers defined in previous
tests end up being visible by LTTng.

Ease up the passing requirement for a test so that it can
still work in these conditions.

Signed-off-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>