Adapt tests to babeltrace 2 output.
The application context type is not present when using babeltrace2 as of
today. This change might get reversed as needed in the future if the
omission of the type was an error on the reader side. For now remove
the type from the lookup and validate based on value only.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>